Simplify Django settings files
