+++ OIPA/OIPA/development_settings.py
@@ -1,57 +1,22 @@
-# Sample production settings, change as needed
+from OIPA.settings import *  # noqa: F401, F403
 
-from OIPA.base_settings import *
-
-DEBUG = True
-FTS_ENABLED = True
-
-MIDDLEWARE_CLASSES += [
+MIDDLEWARE_CLASSES += [  # noqa: F405
     'debug_toolbar.middleware.DebugToolbarMiddleware',
     'yet_another_django_profiler.middleware.ProfilerMiddleware'
 ]
 
-INSTALLED_APPS += {
+INSTALLED_APPS += {  # noqa: F405
     'debug_toolbar',
     'yet_another_django_profiler'
 }
 
+
 def custom_show_toolbar(self):
     return True
+
 
 SECRET_KEY = '__DEV_SECRET_KEY__'
 
 DEBUG_TOOLBAR_CONFIG = {
     'SHOW_TOOLBAR_CALLBACK': custom_show_toolbar,
 }
-
-DATABASES = {
-    'default': {
-        'ENGINE': 'django.contrib.gis.db.backends.postgis',
-        'NAME': 'oipa',
-        'USER': 'oipa',
-        'PASSWORD': 'oipa',
-        'HOST': '127.0.0.1',
-    },
-}
-
-MEDIA_ROOT = os.path.join(BASE_DIR, "media")
-STATIC_ROOT = os.path.join(BASE_DIR, 'static_served/')
-
-# Additional locations of static files
-STATICFILES_DIRS = (
-     os.path.join(BASE_DIR, 'static/'),
-)
-
-FIXTURE_DIRS = (
-     os.path.join(BASE_DIR, '../fixtures/'),
-)
-
-CKAN_URL = "https://iati-staging.ckan.io"
-
-API_CACHE_SECONDS = 0
-
-
-try:
-    from local_settings import *
-except ImportError:
-    pass
