+++ OIPA/OIPA/travis_db_settings.py
@@ -1,4 +1,4 @@
-from OIPA.production_settings import *
+from OIPA.production_settings import *  # noqa: F401, F403
 
 SECRET_KEY = 'j!bxt0h-=d)1@2r8du!+e4m9x-y*5od7+zq&=tfjwq(ecuov!*'
 
@@ -8,5 +8,5 @@ DATABASES = {
         'NAME': 'postgres',
         'USER': 'postgres',
         'PASSWORD': '',
-    },    
+    }
 }
