+++ OIPA/OIPA/travis_test_settings.py
@@ -1,10 +1,10 @@
-from OIPA.production_settings import *
+from OIPA.production_settings import *  # noqa: F401, F403
 
 DATABASES = {
     'default': {
         'ENGINE': 'django.contrib.gis.db.backends.spatialite',
         'NAME': ':memory:',
-    },    
+    },
 }
 
 FTS_ENABLED = False
